fix(#6181): risk 报告方法 Decimal 算术混算修复（liquidity/volatility）

+++ src/ginkgo/trading/risk_management/volatility_risk.py
@@ -253,7 +253,11 @@ class VolatilityRisk(BaseRiskManagement):
         weighted_volatility = 0.0
         for code, position in positions.items():
             if position and position.market_value > 0:
-                weight = position.market_value / total_value
+                # #6181: market_value/total_value 均为 Decimal（ADR-010，自 1803ce39），
+                # weight 是 Decimal；stock_volatility 为 float。Decimal × float 抛
+                # TypeError。weight 转 float（权重为无量纲比值，float 无失真风险），
+                # 对齐方法签名 -> float。对齐 #6180 / #6181 修法。
+                weight = float(position.market_value / total_value)
                 stock_volatility = self._get_stock_volatility(code)
                 weighted_volatility += weight * stock_volatility
